value of rating change to float beacuse of point rating and error of 404 is handled by nocontent ();

+++ src/main/java/com/rating/rating/model/Rating.java
@@ -46,7 +46,7 @@ public class Rating {
     
     @NonNull
     private Long appointmentId;
-    private Long value ;
+    private float value ;
     private String note;
